+++ server/server_unit_test.go
@@ -1,4 +1,4 @@
-package docker
+package server
 
 import (
 	"github.com/dotcloud/docker/utils"
